Update Widget editor
