+++ spyder/app/mainwindow.py
@@ -1231,9 +1231,9 @@ class MainWindow(QMainWindow):
 
         # Hide Internal Console so that people don't use it instead of
         # the External or IPython ones
-        if self.console.dockwidget.isVisible() and DEV is None:
-            self.console.toggle_view_action.setChecked(False)
-            self.console.dockwidget.hide()
+        #if self.console.dockwidget.isVisible() and DEV is None:
+        #    self.console.toggle_view_action.setChecked(False)
+        #    self.console.dockwidget.hide()
 
         # Show Help and Consoles by default
         plugins_to_show = []
